+++ test/test_ops/test_exp.py
@@ -18,3 +18,20 @@ def test_exp():
     z.backward()
 
     assert np.allclose(zt.cpu(), z.detach())
+
+def test_exp_grad():
+    x = torch.empty((5, 10)).random_(10).requires_grad_()
+    e = x.exp()
+    z = e.sum()
+    e.retain_grad()
+    z.retain_grad()
+    z.backward()
+
+    xt = Tensor(x.detach().numpy())
+    et = xt.exp()
+    zt = et.sum()
+    zt.backward()
+
+    assert np.allclose(x.detach().numpy(), xt.data)
+    assert np.allclose(e.detach().numpy(), et.data)
+    assert np.allclose(e.grad, et.grad.data)
